test create git propertites
