+++ package.json
@@ -23,7 +23,7 @@
     "clean": "rimraf lib coverage *.tgz",
     "serve": "nodemon _example/index.js --watch _example --watch src",
     "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold\" \"npm run watch-ts\" \"npm run serve\"",
-    "test": "jest --forceExit",
+    "test": "jest --forceExit --detectOpenHandles",
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watchAll",
     "build-ts": "tsc",
